requirements.txt
# Dependências do dashboard (Streamlit)
pandas<2.0
numpy<1.24
requests==2.31.0
streamlit
scikit-learn==1.7.1
matplotlib
seaborn
plotly

# Dependências da função de coleta (Vercel)
psycopg2-binary==2.9.7



